+++ auto_route/example/lib/mobile/main.dart
@@ -1,4 +1,4 @@
-import 'package:auto_route/auto_route.dart';
+import 'package:example/data/db.dart';
 import 'package:example/mobile/router/auth_guard.dart';
 import 'package:example/mobile/router/router.gr.dart';
 import 'package:flutter/cupertino.dart';
@@ -6,19 +6,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 
-import '../data/db.dart';
-
 void main() => runApp(MyApp());
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
+  final authService = AuthService();
+
   final _rootRouter = RootRouter(
       // authGuard: AuthGuard(),
-
       );
 
   @override
@@ -30,7 +29,7 @@ class _MyAppState extends State<MyApp> {
       routeInformationParser: _rootRouter.defaultRouteParser(),
       builder: (_, router) {
         return ChangeNotifierProvider<AuthService>(
-          create: (_) => AuthService(),
+          create: (_) => authService,
           child: BooksDBProvider(
             child: router!,
           ),
@@ -38,9 +37,4 @@ class _MyAppState extends State<MyApp> {
       },
     );
   }
-}
-
-class MyObserver extends AutoRouterObserver {
-  @override
-  void didPush(Route route, Route? previousRoute) {}
 }